Move PEP 723 utilities to req
