+++ src/pip/_internal/req/pep723.py
@@ -21,15 +21,21 @@ def pep723_metadata(scriptfile: str) -> dict[str, Any]:
     matches = list(
         filter(lambda m: m.group("type") == name, re.finditer(REGEX, script))
     )
+
     if len(matches) > 1:
-        raise PEP723Exception(f"Multiple {name} blocks found")
+        raise PEP723Exception(f"Multiple {name!r} blocks found in {scriptfile!r}")
     elif len(matches) == 1:
         content = "".join(
             line[2:] if line.startswith("# ") else line[1:]
             for line in matches[0].group("content").splitlines(keepends=True)
         )
-        return tomllib.loads(content)
+        try:
+            metadata = tomllib.loads(content)
+        except Exception as exc:
+            raise PEP723Exception(f"Failed to parse TOML in {scriptfile!r}") from exc
     else:
         raise PEP723Exception(
-            f"File does not contain '{name}' metadata: {scriptfile!r}"
+            f"File does not contain {name!r} metadata: {scriptfile!r}"
         )
+
+    return metadata
